+++ examples/benchmark_router_round_robin.rs
@@ -1,7 +1,7 @@
 use serde::Serialize;
 use std::error::Error;
 use std::process::exit;
-use std::time::{Duration, Instant};
+use std::time::Instant;
 use tyra::prelude::*;
 use tyra::router::{AddActorMessage, RoundRobinRouterFactory};
 
@@ -138,7 +138,7 @@ impl Handler<Finish> for Aggregator {
                 "{} It took {:?} to finish {} actors",
                 self.name, duration, self.total_actors
             );
-            self.ctx.system.stop(Duration::from_secs(60));
+            self.ctx.system.stop();
         }
         Ok(ActorResult::Ok)
     }