+++ examples/benchmark_single_actor.rs
@@ -1,7 +1,7 @@
 use serde::Serialize;
 use std::error::Error;
 use std::process::exit;
-use std::time::{Duration, Instant};
+use std::time::Instant;
 use tyra::prelude::*;
 
 #[derive(Hash, Serialize)]
@@ -63,7 +63,7 @@ impl Handler<MessageA> for Benchmark {
             );
         }
         if self.count == self.total_msgs {
-            context.system.stop(Duration::from_secs(60));
+            context.system.stop();
         }
         Ok(ActorResult::Ok)
     }