+++ src/system/system_state.rs
@@ -71,7 +71,7 @@ impl SystemState {
         let now = Instant::now();
 
         while self.get_actor_count() != 0 {
-            if now.elapsed() >= timeout || self.is_force_stopped.load(Ordering::Relaxed) {
+            if (timeout.as_secs() > 0 && now.elapsed() >= timeout) || self.is_force_stopped.load(Ordering::Relaxed) {
                 self.is_force_stopped.store(true, Ordering::Relaxed);
                 self.mailboxes.clear();
                 break;